Add ability to store multiple sleeperIds on a user
This change updates the model to separate out the sleeperOwnerId to its
own table, which lets us have multiple owner IDs on a single user.

This is still a manual process to be done by the admins, but at least
this lets us account for people with multiple sleeper accounts.

This change doesn't account for people having multiple discord accounts,
we probably need a merge tool of some sort in the future but for now,
it's so much of an edge case that it doesn't matter too much.

Closes #56