Fixes spread pool reset issue
This fixes an issue with the reset button not actually applying for a
bet that was already entered. This is caused by a reset not making the
component render. Easiest way to solve it was to keep track of if the
reset button was hit recently and use a hidden input component with the
team that was the previous bet so that it makes sure to zero it out.

This also does a minor optimization to not call handleChange twice. It
also sets the default to show the slider if an existing bet is in the
system, since we'd want that bet to be displayed anyhow.