Add the concept of a season week to the database
This commit adds the SeasonWeek to a season, which is useful for keeping
track of what week we're currently in, which will be nice for displaying
results of things and toggle things over correctly.

While this isn't the solution for #72, this is probably something we
could look at to model the actual solution for #72.